Replace frontend and update backend security alerts

+++ src/main/java/com/sentinelcore/repository/AssetRepository.java
@@ -3,6 +3,6 @@ package com.sentinelcore.repository;
 import com.sentinelcore.Asset;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AssetRepository extends JpaRepository<Asset, Integer> {
+public interface AssetRepository extends JpaRepository<Asset, Long> {
 
 }
